Move HREXPlots for consistency
